requirements.txt
requests
certifi
confluent-kafka[avro,json,protobuf]>=1.4.2
python-dotenv
